同步一波内容

+++ src/main/java/cn/nukkit/entity/projectile/EntityArrow.java
@@ -86,11 +86,11 @@ public class EntityArrow extends EntityProjectile {
     }
 
     @Override
-    public float getResultDamage() {
-        float base = super.getResultDamage();
+    public int getResultDamage() {
+        int base = super.getResultDamage();
 
         if (this.isCritical()) {
-            base += ThreadLocalRandom.current().nextInt((int) (base / 2) + 2);
+            base += ThreadLocalRandom.current().nextInt(base / 2 + 2);
         }
 
         return base;
